pyproject.toml
[tool.poetry]
name = "book_api"
version = "0.1.0"
description = ""
authors = ["fathibensari <fethibensari@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
pip = "^22.3"
install = "^1.3.5"
djangorestframework-queryfields = "^1.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.0"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
